Refactor: Reconnect updated module. shell. #107
